Renames from artin_representation/databases/ to typed_data/
