+++ lmfdb/typed_data/Lfunction_types.py
