+++ lmfdb/typed_data/__init__.py
@@ -1,9 +1,9 @@
 from lmfdb.utils import make_logger
-databases_logger = make_logger("databases_logger", hl=True)
+typed_data_logger = make_logger("typed_data_logger", hl=True)
 
 import type_generation
 assert type_generation
 import standard_types
 assert standard_types
-import Dokchitser_databases
-assert Dokchitser_databases
+import artin_types
+assert artin_types
