+++ lmfdb/typed_data/artin_types.py
